Keep grouped weight quantizers in DeepGEMM module
